requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
passlib[bcrypt]
python-jose[cryptography]
python-multipart